Move everything related to actually decoding text to a new module
